+++ lib/http/intercepted_client.dart
@@ -3,42 +3,50 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:http/http.dart';
-import 'package:http_interceptor/models/models.dart';
 import 'package:http_interceptor/extensions/extensions.dart';
+import 'package:http_interceptor/models/models.dart';
 
 import 'http_methods.dart';
 import 'interceptor_contract.dart';
 
-///Class to be used by the user to set up a new `http.Client` with interceptor supported.
-///call the `build()` constructor passing in the list of interceptors.
-///Example:
-///```dart
-/// InterceptedClient httpClient = InterceptedClient.build(interceptors: [
-///     Logger(),
-/// ]);
-///```
+/// Class to be used by the user to set up a new `http.Client` with interceptor
+/// support.
 ///
-///Then call the functions you want to, on the created `httpClient` object.
-///```dart
-/// httpClient.get(...);
-/// httpClient.post(...);
-/// httpClient.put(...);
-/// httpClient.delete(...);
-/// httpClient.head(...);
-/// httpClient.patch(...);
-/// httpClient.read(...);
-/// httpClient.readBytes(...);
-/// httpClient.close();
-///```
-///Don't forget to close the client once you are done, as a client keeps
-///the connection alive with the server.
+/// Call `build()` and pass list of interceptors as parameter.
 ///
-///Note: `send` method is not currently supported.
+/// Example:
+/// ```dart
+///  InterceptedClient client = InterceptedClient.build(interceptors: [
+///      LoggingInterceptor(),
+///  ]);
+/// ```
+///
+/// Then call the functions you want to, on the created `client` object.
+/// ```dart
+///  client.get(...);
+///  client.post(...);
+///  client.put(...);
+///  client.delete(...);
+///  client.head(...);
+///  client.patch(...);
+///  client.read(...);
+///  client.send(...);
+///  client.readBytes(...);
+///  client.close();
+/// ```
+///
+/// Don't forget to close the client once you are done, as a client keeps
+/// the connection alive with the server by default.
 class InterceptedClient extends BaseClient {
-  List<InterceptorContract> interceptors;
-  Duration? requestTimeout;
-  RetryPolicy? retryPolicy;
-  String Function(Uri)? findProxy;
+  /// List of interceptors that will be applied to the requests and responses.
+  final List<InterceptorContract> interceptors;
+
+  /// Maximum duration of a request.
+  final Duration? requestTimeout;
+
+  /// A policy that defines whether a request or response should trigger a
+  /// retry. This is useful for implementing JWT token expiration
+  final RetryPolicy? retryPolicy;
 
   int _retryCount = 0;
   late Client _inner;
@@ -47,22 +55,36 @@ class InterceptedClient extends BaseClient {
     required this.interceptors,
     this.requestTimeout,
     this.retryPolicy,
-    this.findProxy,
     Client? client,
   }) : _inner = client ?? Client();
 
+  /// Builds a new [InterceptedClient] instance.
+  ///
+  /// Interceptors are applied in a linear order. For example a list that looks
+  /// like this:
+  ///
+  /// ```dart
+  /// InterceptedClient.build(
+  ///   interceptors: [
+  ///     WeatherApiInterceptor(),
+  ///     LoggerInterceptor(),
+  ///   ],
+  /// ),
+  /// ```
+  ///
+  /// Will apply first the `WeatherApiInterceptor` interceptor, so when
+  /// `LoggerInterceptor` receives the request/response it has already been
+  /// intercepted.
   factory InterceptedClient.build({
     required List<InterceptorContract> interceptors,
     Duration? requestTimeout,
     RetryPolicy? retryPolicy,
-    String Function(Uri)? findProxy,
     Client? client,
   }) =>
       InterceptedClient._internal(
         interceptors: interceptors,
         requestTimeout: requestTimeout,
         retryPolicy: retryPolicy,
-        findProxy: findProxy,
         client: client,
       );
 
@@ -70,24 +92,24 @@ class InterceptedClient extends BaseClient {
   Future<Response> head(
     Uri url, {
     Map<String, String>? headers,
-  }) =>
-      _sendUnstreamed(
-        method: Method.HEAD,
+  }) async =>
+      (await _sendUnstreamed(
+        method: HttpMethod.HEAD,
         url: url,
         headers: headers,
-      );
+      )) as Response;
 
   Future<Response> get(
     Uri url, {
     Map<String, String>? headers,
     Map<String, dynamic>? params,
-  }) =>
-      _sendUnstreamed(
-        method: Method.GET,
+  }) async =>
+      (await _sendUnstreamed(
+        method: HttpMethod.GET,
         url: url,
         headers: headers,
         params: params,
-      );
+      )) as Response;
 
   @override
   Future<Response> post(
@@ -96,15 +118,15 @@ class InterceptedClient extends BaseClient {
     Map<String, dynamic>? params,
     Object? body,
     Encoding? encoding,
-  }) =>
-      _sendUnstreamed(
-        method: Method.POST,
+  }) async =>
+      (await _sendUnstreamed(
+        method: HttpMethod.POST,
         url: url,
         headers: headers,
         params: params,
         body: body,
         encoding: encoding,
-      );
+      )) as Response;
 
   @override
   Future<Response> put(
@@ -113,15 +135,15 @@ class InterceptedClient extends BaseClient {
     Map<String, dynamic>? params,
     Object? body,
     Encoding? encoding,
-  }) =>
-      _sendUnstreamed(
-        method: Method.PUT,
+  }) async =>
+      (await _sendUnstreamed(
+        method: HttpMethod.PUT,
         url: url,
         headers: headers,
         params: params,
         body: body,
         encoding: encoding,
-      );
+      )) as Response;
 
   @override
   Future<Response> patch(
@@ -130,15 +152,15 @@ class InterceptedClient extends BaseClient {
     Map<String, dynamic>? params,
     Object? body,
     Encoding? encoding,
-  }) =>
-      _sendUnstreamed(
-        method: Method.PATCH,
+  }) async =>
+      (await _sendUnstreamed(
+        method: HttpMethod.PATCH,
         url: url,
         headers: headers,
         params: params,
         body: body,
         encoding: encoding,
-      );
+      )) as Response;
 
   @override
   Future<Response> delete(
@@ -147,15 +169,15 @@ class InterceptedClient extends BaseClient {
     Map<String, dynamic>? params,
     Object? body,
     Encoding? encoding,
-  }) =>
-      _sendUnstreamed(
-        method: Method.DELETE,
+  }) async =>
+      (await _sendUnstreamed(
+        method: HttpMethod.DELETE,
         url: url,
         headers: headers,
         params: params,
         body: body,
         encoding: encoding,
-      );
+      )) as Response;
 
   @override
   Future<String> read(
@@ -181,14 +203,19 @@ class InterceptedClient extends BaseClient {
     });
   }
 
-  // TODO(codingalecr): Implement interception from `send` method.
   @override
-  Future<StreamedResponse> send(BaseRequest request) {
-    return _inner.send(request);
+  Future<StreamedResponse> send(BaseRequest request) async {
+    final interceptedRequest = await _interceptRequest(request);
+
+    final response = await _inner.send(interceptedRequest);
+
+    final interceptedResponse = await _interceptResponse(response);
+
+    return interceptedResponse as StreamedResponse;
   }
 
-  Future<Response> _sendUnstreamed({
-    required Method method,
+  Future<BaseResponse> _sendUnstreamed({
+    required HttpMethod method,
     required Uri url,
     Map<String, String>? headers,
     Map<String, dynamic>? params,
@@ -197,7 +224,7 @@ class InterceptedClient extends BaseClient {
   }) async {
     url = url.addParameters(params);
 
-    Request request = new Request(methodToString(method), url);
+    Request request = new Request(method.asString, url);
     if (headers != null) request.headers.addAll(headers);
     if (encoding != null) request.encoding = encoding;
     if (body != null) {
@@ -231,21 +258,22 @@ class InterceptedClient extends BaseClient {
 
   /// Attempts to perform the request and intercept the data
   /// of the response
-  Future<Response> _attemptRequest(Request request) async {
-    var response;
+  Future<BaseResponse> _attemptRequest(BaseRequest request) async {
+    BaseResponse response;
     try {
       // Intercept request
       final interceptedRequest = await _interceptRequest(request);
 
       var stream = requestTimeout == null
-          ? await send(interceptedRequest)
-          : await send(interceptedRequest).timeout(requestTimeout!);
+          ? await _inner.send(interceptedRequest)
+          : await _inner.send(interceptedRequest).timeout(requestTimeout!);
 
-      response = await Response.fromStream(stream);
+      response =
+          request is Request ? await Response.fromStream(stream) : stream;
+
       if (retryPolicy != null &&
           retryPolicy!.maxRetryAttempts > _retryCount &&
-          await retryPolicy!.shouldAttemptRetryOnResponse(
-              ResponseData.fromHttpResponse(response))) {
+          await retryPolicy!.shouldAttemptRetryOnResponse(response)) {
         _retryCount += 1;
         return _attemptRequest(request);
       }
@@ -265,27 +293,27 @@ class InterceptedClient extends BaseClient {
   }
 
   /// This internal function intercepts the request.
-  Future<Request> _interceptRequest(Request request) async {
+  Future<BaseRequest> _interceptRequest(BaseRequest request) async {
+    BaseRequest interceptedRequest = request.copyWith();
     for (InterceptorContract interceptor in interceptors) {
-      RequestData interceptedData = await interceptor.interceptRequest(
-        data: RequestData.fromHttpRequest(request),
+      interceptedRequest = await interceptor.interceptRequest(
+        request: interceptedRequest,
       );
-      request = interceptedData.toHttpRequest();
     }
 
-    return request;
+    return interceptedRequest;
   }
 
   /// This internal function intercepts the response.
-  Future<Response> _interceptResponse(Response response) async {
+  Future<BaseResponse> _interceptResponse(BaseResponse response) async {
+    BaseResponse interceptedResponse = response;
     for (InterceptorContract interceptor in interceptors) {
-      ResponseData responseData = await interceptor.interceptResponse(
-        data: ResponseData.fromHttpResponse(response),
+      interceptedResponse = await interceptor.interceptResponse(
+        response: interceptedResponse,
       );
-      response = responseData.toHttpResponse();
     }
 
-    return response;
+    return interceptedResponse;
   }
 
   void close() {