+++ lib/http/intercepted_http.dart
@@ -1,61 +1,100 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
+
 import 'package:http/http.dart';
 import 'package:http_interceptor/http_interceptor.dart';
 
 import 'intercepted_client.dart';
 import 'interceptor_contract.dart';
 
-///Class to be used by the user as a replacement for 'http' with interceptor supported.
-///call the `build()` constructor passing in the list of interceptors.
-///Example:
-///```dart
-/// InterceptedHttp http = InterceptedHttp.build(interceptors: [
-///     Logger(),
-/// ]);
-///```
-///Then call the functions you want to, on the created `http` object.
-///```dart
-/// http.get(...);
-/// http.post(...);
-/// http.put(...);
-/// http.delete(...);
-/// http.head(...);
-/// http.patch(...);
-/// http.read(...);
-/// http.readBytes(...);
-///```
+/// Class to be used by the user as a replacement for 'http' with interceptor
+/// support.
+///
+/// It is a useful class if you want to centralize HTTP request calls
+/// since it creates and discards [InterceptedClient] instances after the
+/// request is done and that allows you to avoid handling your own [Client]
+/// instances.
+///
+///
+/// Call `build()` and pass list of interceptors as parameter.
+///
+/// Example:
+/// ```dart
+///  InterceptedHttp http = InterceptedHttp.build(interceptors: [
+///      LoggingInterceptor(),
+///  ]);
+/// ```
+///
+/// Then call the functions you want to, on the created `http` object.
+/// ```dart
+///  http.get(...);
+///  http.post(...);
+///  http.put(...);
+///  http.delete(...);
+///  http.head(...);
+///  http.patch(...);
+///  http.send(...);
+///  http.read(...);
+///  http.readBytes(...);
+/// ```
 class InterceptedHttp {
-  List<InterceptorContract> interceptors;
-  Duration? requestTimeout;
-  RetryPolicy? retryPolicy;
-  String Function(Uri)? findProxy;
+  /// List of interceptors that will be applied to the requests and responses.
+  final List<InterceptorContract> interceptors;
+
+  /// Maximum duration of a request.
+  final Duration? requestTimeout;
+
+  /// A policy that defines whether a request or response should trigger a
+  /// retry. This is useful for implementing JWT token expiration
+  final RetryPolicy? retryPolicy;
+
+  /// Inner client that is wrapped for intercepting.
+  ///
+  /// If you don't specify your own client then the library will instantiate
+  /// a default one.
   Client? client;
 
   InterceptedHttp._internal({
     required this.interceptors,
     this.requestTimeout,
     this.retryPolicy,
-    this.findProxy,
     this.client,
   });
 
+  /// Builds a new [InterceptedHttp] instance. It helps avoid creating and
+  /// managing your own `Client` instances.
+  ///
+  /// Interceptors are applied in a linear order. For example a list that looks
+  /// like this:
+  ///
+  /// ```dart
+  /// InterceptedClient.build(
+  ///   interceptors: [
+  ///     WeatherApiInterceptor(),
+  ///     LoggerInterceptor(),
+  ///   ],
+  /// ),
+  /// ```
+  ///
+  /// Will apply first the `WeatherApiInterceptor` interceptor, so when
+  /// `LoggerInterceptor` receives the request/response it has already been
+  /// intercepted.
   factory InterceptedHttp.build({
     required List<InterceptorContract> interceptors,
     Duration? requestTimeout,
     RetryPolicy? retryPolicy,
-    String Function(Uri)? findProxy,
     Client? client,
   }) =>
       InterceptedHttp._internal(
         interceptors: interceptors,
         requestTimeout: requestTimeout,
         retryPolicy: retryPolicy,
-        findProxy: findProxy,
         client: client,
       );
 
+  /// Performs a HEAD request with a new [Client] instance and closes it after
+  /// it has been used.
   Future<Response> head(
     url, {
     Map<String, String>? headers,
@@ -66,6 +105,8 @@ class InterceptedHttp {
         ));
   }
 
+  /// Performs a GET request with a new [Client] instance and closes it after
+  /// it has been used.
   Future<Response> get(
     url, {
     Map<String, String>? headers,
@@ -78,6 +119,8 @@ class InterceptedHttp {
         ));
   }
 
+  /// Performs a POST request with a new [Client] instance and closes it after
+  /// it has been used.
   Future<Response> post(
     url, {
     Map<String, String>? headers,
@@ -94,6 +137,8 @@ class InterceptedHttp {
         ));
   }
 
+  /// Performs a PUT request with a new [Client] instance and closes it after
+  /// it has been used.
   Future<Response> put(
     url, {
     Map<String, String>? headers,
@@ -110,6 +155,8 @@ class InterceptedHttp {
         ));
   }
 
+  /// Performs a PATCH request with a new [Client] instance and closes it after
+  /// it has been used.
   Future<Response> patch(
     url, {
     Map<String, String>? headers,
@@ -126,6 +173,8 @@ class InterceptedHttp {
         ));
   }
 
+  /// Performs a DELETE request with a new [Client] instance and closes it after
+  /// it has been used.
   Future<Response> delete(
     url, {
     Map<String, String>? headers,
@@ -142,6 +191,8 @@ class InterceptedHttp {
         ));
   }
 
+  /// Executes `client.read` with a new [Client] instance and closes it after
+  /// it has been used.
   Future<String> read(
     url, {
     Map<String, String>? headers,
@@ -154,6 +205,8 @@ class InterceptedHttp {
         ));
   }
 
+  /// Executes `client.readBytes` with a new [Client] instance and closes it
+  /// after it has been used.
   Future<Uint8List> readBytes(
     url, {
     Map<String, String>? headers,
@@ -165,6 +218,8 @@ class InterceptedHttp {
             params: params,
           ));
 
+  /// Internal convenience utility to create a new [Client] instance for each
+  /// request. It closes the client after using it for the request.
   Future<T> _withClient<T>(
     Future<T> fn(InterceptedClient client),
   ) async {
@@ -172,7 +227,6 @@ class InterceptedHttp {
       interceptors: interceptors,
       requestTimeout: requestTimeout,
       retryPolicy: retryPolicy,
-      findProxy: findProxy,
       client: this.client,
     );
     try {